requirements.txt
streamlit==1.28.1
google-generativeai
textstat==0.7.3
python-dotenv==1.0.0
